+++ packages/tgweb/lib/tgweb/create_initially.mjs
@@ -1,13 +1,15 @@
 import * as PATH from "path"
+import { slash } from "./slash.mjs"
 import fs from "fs"
 import getType from "./get_type.mjs"
 import generateHTML from "./generate_html.mjs"
 
 const createInitially = (path, siteData) => {
-  const dirname = PATH.dirname(path)
+  const dirname = slash(PATH.dirname(path))
 
   if (dirname.startsWith("src/images") || dirname.startsWith("src/audios")) {
-    const targetPath = path.replace(/^src\//, "dist/")
+    const distPath = slash(path).replace(/^src\//, "dist/")
+    const targetPath = PATH.resolve(distPath)
     const targetDir = PATH.dirname(targetPath)
 
     if (!fs.existsSync(targetDir)) fs.mkdirSync(targetDir, { recursive: true })
@@ -21,13 +23,14 @@ const createInitially = (path, siteData) => {
     const html = generateHTML(path, siteData)
 
     if (html !== undefined) {
-      const targetPath = path.replace(/^src\//, "dist/").replace(/^dist\/pages\//, "dist/")
+      const distPath = slash(path).replace(/^src\//, "dist/").replace(/^dist\/pages\//, "dist/")
+      const targetPath = PATH.resolve(distPath)
       const targetDir = PATH.dirname(targetPath)
 
       if (!fs.existsSync(targetDir)) fs.mkdirSync(targetDir, { recursive: true })
       fs.writeFileSync(targetPath, html)
 
-      if (process.env.VERBOSE) console.log(`Created ${path}.`)
+      if (process.env.VERBOSE) console.log(`Created ${distPath}.`)
     }
   }
 }