+++ packages/tgweb/lib/tgweb/generate_html.mjs
@@ -1,3 +1,4 @@
+import { slash } from "./slash.mjs"
 import getType from "./get_type.mjs"
 import { renderPage } from "./render_page.mjs"
 import { renderArticle } from "./render_article.mjs"
@@ -5,8 +6,8 @@ import { renderArticle } from "./render_article.mjs"
 const generateHTML = (path, siteData) => {
   const type = getType(path)
 
-  if (type === "page") return renderPage(path, siteData)
-  else if (type === "article") return renderArticle(path, siteData)
+  if (type === "page") return renderPage(slash(path), siteData)
+  else if (type === "article") return renderArticle(slash(path), siteData)
 }
 
 export default generateHTML