+++ packages/tgweb/lib/tgweb/get_wrapper.mjs
@@ -1,12 +1,13 @@
 import * as PATH from "path"
+import { slash } from "./slash.mjs"
 
 const getWrapper = (siteData, path) => {
-  const dirParts = path.split(PATH.sep)
+  const dirParts = slash(path).split(PATH.posix.sep)
   dirParts.pop()
 
   for(let i = dirParts.length; i > 0; i--) {
-    const dir = dirParts.slice(0, i).join(PATH.sep)
-    const wrapperPath = PATH.join(dir, "_wrapper.html")
+    const dir = dirParts.slice(0, i).join(PATH.posix.sep)
+    const wrapperPath = slash(PATH.join(dir, "_wrapper.html"))
 
     const wrapper = siteData.wrappers.find(wrapper => wrapper.path === wrapperPath)
     if (wrapper) return wrapper