+++ packages/tgweb/lib/tgweb/slash.mjs
@@ -1,0 +1,5 @@
+import * as PATH from "path"
+
+const slash = path => path.split(PATH.sep).join(PATH.posix.sep)
+
+export { slash }