+++ packages/tgweb/test/tgweb/get_site_data.test.mjs
@@ -80,7 +80,7 @@ describe("getSiteData", () => {
     assert.equal(siteData.articles.length, 8)
   })
 
-  it("should make the article front matter inherit site and wrapper properties", () => {
+  it("should make the article front matter inherit site, layout and wrapper properties", () => {
     const wd = PATH.resolve(__dirname, "../examples/site_1")
     const siteData = getSiteData(wd)
 
@@ -91,6 +91,10 @@ describe("getSiteData", () => {
     assert.equal(article.frontMatter["data-current-year"], 2023)
     assert.equal(article.frontMatter["property-og:image"],
       "http://localhost:3000/images/red_square.png")
+    assert.equal(article.frontMatter["data-w"], 0)
+    assert.equal(article.frontMatter["data-x"], 1)
+    assert.equal(article.frontMatter["data-y"], 2)
+    assert.equal(article.frontMatter["data-z"], 3)
   })
 
   it("should return the site data with dependencies", () => {